update buildTest for type action
- split up buildTest into separate files
