+++ packages/build-test/tests/formatMethod.test.ts
@@ -1,0 +1,9 @@
+import { formatMethod } from "../src/formatMethod";
+
+describe("formatMethod", () => {
+  it("throws error if invalid step action", () => {
+    expect(() => {
+      formatMethod("drag" as any, 0);
+    }).toThrowError();
+  });
+});
